+++ src/react-loader.js
@@ -1,6 +1,6 @@
-const React = require('react'); //eslint-disable-line no-unused-vars
-const deepEqual = require('deep-eql');
-const isPromise = require('is-promise');
+import React from 'react'; //eslint-disable-line no-unused-vars
+import deepEqual from 'deep-eql';
+import isPromise from 'is-promise';
 
 const STATUS_FETCHING = 0;
 const STATUS_LOADED = 1;
@@ -11,29 +11,24 @@ const NODE_ENV = process.env.NODE_ENV || 'production';
 const debug = (NODE_ENV === 'development') ? console.log : () => {};
 debug('Environment:', NODE_ENV);
 
-module.exports = function ReactLoader(_options = {}) {
+export default function ReactLoader(_options = {}) {
 	const options = Object.assign({
-		component: null,
 		load: null,
 		resultProp: 'data',
+		errorProp: 'error',
 		errorComponent: () => (<div>Impossible to fetch the data requested.</div>),
 		loadingComponent: () => (<div>Loading...</div>),
 		shouldComponentReload: (props, nextProps) => !deepEqual(props, nextProps),
 		componentWillUnmount: () => {},
 	}, _options);
 
-	if (!options.component) {
-		throw new Error('ReactLoader : No component defined. Cannot create');
-	}
-	const displayName = `ReactLoader(${options.component.displayName || options.component.name})`;
 	if (!options.load) {
-		throw new Error(`${displayName} : No load() defined. Cannot create`);
+		throw new Error('ReactLoader : No load() defined. Cannot create');
 	}
 	if (typeof(options.load) !== 'function') {
-		throw new Error(`${displayName} : load must be a function returning a Promise. Cannot create`);
+		throw new Error('ReactLoader : load must be a function returning a Promise. Cannot create');
 	}
 
-	const Component = options.component;
 	const ErrorComponent = options.errorComponent;
 	const LoadingComponent = options.loadingComponent;
 
@@ -41,8 +36,9 @@ module.exports = function ReactLoader(_options = {}) {
 	const componentWillUnmount = options.componentWillUnmount;
 	const shouldComponentReload = options.shouldComponentReload;
 	const resultProp = options.resultProp;
+	const errorProp = options.errorProp;
 
-	class Loader extends React.Component {
+	class _Loader extends React.Component {
 		constructor(props) {
 			super(props);
 			this.state = {
@@ -53,7 +49,9 @@ module.exports = function ReactLoader(_options = {}) {
 		componentDidMount() {
 			this.doLoad();
 		}
-		componentDidUpdate(prevProps) {
+
+		//eslint-disable-next-line no-unused-vars
+		componentDidUpdate(prevProps, prevState, snapshot) {
 			if (shouldComponentReload(prevProps, this.props)) {
 				this.doLoad();
 			}
@@ -64,7 +62,7 @@ module.exports = function ReactLoader(_options = {}) {
 			this.setState({status: STATUS_FETCHING});
 			const prom = load(this.props);
 			if (!isPromise(prom)) {
-				throw new TypeError(`${displayName} : load(props) must return a Promise/A+ compliant object`);
+				throw new TypeError('ReactLoader : load(props) must return a Promise/A+ compliant object');
 			}
 
 			prom.then((data) => {
@@ -89,22 +87,13 @@ module.exports = function ReactLoader(_options = {}) {
 
 			if (this.state.status === STATUS_ERROR) {
 				const forwardProps = Object.assign({}, props);
-				if (resultProp) {
-					forwardProps[resultProp] = this.state.error;
-				}
+				forwardProps[errorProp] = this.state.error;
 				return (
 					<ErrorComponent {...forwardProps}/>
 				);
 			}
 			else if (this.state.status === STATUS_LOADED) {
-				const forwardProps = Object.assign({}, props);
-				if (resultProp) {
-					forwardProps[resultProp] = this.state.data;
-				}
-				debug(props, ' => ', forwardProps);
-				return (
-					<Component {...forwardProps}/>
-				);
+				return this.renderComponent();
 			}
 			else {
 				return (
@@ -113,12 +102,35 @@ module.exports = function ReactLoader(_options = {}) {
 			}
 		}
 
-		shouldComponentUpdate(nextProps, nextState) {
+		//eslint-disable-next-line no-unused-vars
+		shouldComponentUpdate(nextProps, nextState, nextContext) {
 			return !deepEqual(this.props, nextProps) || !deepEqual(this.state, nextState);
 		}
 	}
-	Loader.displayName = displayName;
-	Loader.propTypes = Component.propTypes;
 
-	return Loader;
-};
+	const outFn = (Component) => {
+		if (typeof(Component)==='object' && Component.kind==='class') {
+			return {
+				...Component,
+				finisher: (Class) => (...args) => {
+					return new (outFn(Class))(...args);
+				},
+			};
+		}
+		else {
+			class Loader extends _Loader {
+				renderComponent() {
+					const forwardProps = Object.assign({}, this.props);
+					forwardProps[resultProp] = this.state.data;
+					return (
+						<Component {...forwardProps}/>
+					);
+				}
+			}
+			Loader.displayName = `ReactLoader(${Component.displayName || Component.name})`;
+			Loader.propTypes = Component.propTypes;
+			return Loader;
+		}
+	};
+	return outFn;
+}