+++ stories/index.stories.js
@@ -14,36 +14,60 @@ storiesOf('ReactLoader')
 
 	//It gets wrapper around the ReactLoader
 	let MyComponent = ReactLoader({
-		component: MyPureComponent,
-		errorComponent: (props) => (<div>An error occured : {JSON.stringify(props.rest_endpoint.error)}</div>),
+		errorComponent: (props) => (<div>An error occurred : {JSON.stringify(props.error)}</div>),
 		loadingComponent: () => (<div>Loading. Content takes 2s to load</div>),
 		load: (props) => {
 			//Faking an async call by waiting 2 seconds
-			return new Promise((resolve, reject) => {
-			    setTimeout(() => resolve(42), 2000)
-			});
+			return Promise.resolve()
+				.then(() => promsleep(2000, 42))
+				.then(() => 42);
 		},
-	});
+	})(MyPureComponent);
 
 	return (
 		<MyComponent/>
 	);
 })
-.add('Full example with error', () => {
+.add('Returns error', () => {
 	//The pure component
 	const MyPureComponent = (props) => (
 		<div>Content loaded : {props.data}</div>
 	);
 
 	const MyComponent = ReactLoader({
-		component: MyPureComponent,
 		resultProp: 'data',
-		errorComponent: (props) => (<div>An error occured : {JSON.stringify(props.data)}</div>),
+		errorComponent: (props) => (<div>An error occurred : {JSON.stringify(props.error)}</div>),
 		loadingComponent: (props) => (<div>Loading. Content takes 2s to load</div>),
 		load: () => {
-			return Promise.resolve().then(() => promsleep(2000, 42)).then(() => { throw 'my_error_message'; });
+			return Promise.resolve()
+				.then(() => promsleep(2000, 42))
+				.then(() => { throw 'my_error_message'; });
 		}
-	});
+	})(MyPureComponent);
+
+	return (
+		<MyComponent/>
+	);
+})
+.add('As a decorator', () => {
+	@ReactLoader({
+		resultProp: 'data',
+		errorComponent: (props) => (<div>An error occurred : {JSON.stringify(props.error)}</div>),
+		loadingComponent: (props) => (<div>Loading. Content takes 2s to load</div>),
+		load: () => {
+			return Promise.resolve()
+				.then(() => promsleep(2000, 42))
+				.then(() => 'OK!');
+		}
+	})
+	class MyComponent extends React.PureComponent {
+		render() {
+			const props = this.props;
+			return (
+				<div>Content loaded : {props.data}</div>
+			);
+		}
+	}
 
 	return (
 		<MyComponent/>
@@ -51,9 +75,9 @@ storiesOf('ReactLoader')
 })
 .add('ERR Missing component', () => {
 	const MyComponent = ReactLoader({
-		componentDidMount: () => {
-			//MyService.asyncCall();
-		}
+		load: () => {
+			return Promise.resolve(42);
+		},
 	});
 
 	return (
@@ -62,8 +86,7 @@ storiesOf('ReactLoader')
 })
 .add('ERR Missing load', () => {
 	const MyComponent = ReactLoader({
-		component: (props) => (<div>Content loaded : {props.data}</div>),
-	});
+	})((props) => (<div>Content loaded : {props.data}</div>));
 
 	return (
 		<MyComponent/>
@@ -71,15 +94,12 @@ storiesOf('ReactLoader')
 })
 .add('ERR load() does not return a promise', () => {
 	const MyComponent = ReactLoader({
-		component: (props) => (<div>Content loaded : {props.data}</div>),
 		load: () => {
 			return 42;
 		}
-	});
+	})((props) => (<div>Content loaded : {props.data}</div>));
 
 	return (
 		<MyComponent/>
 	);
 });
-
-require('./decorators.stories');